no message
git-svn-id: http://svn.beutner.name/trunk/sbnc@254 b3a6972a-4407-0410-a3f0-9bc6c1699d69

+++ Hashtable.h
@@ -63,7 +63,7 @@ public:
 
 	static unsigned char Hash(const char* String) {
 		unsigned char Out = 0;
-		unsigned int Len = strlen(String);
+		size_t Len = strlen(String);
 
 		for (size_t i = 0; i < Len; i++)
 			Out ^= CaseSensitive ? String[i] : toupper(String[i]);
